refactor(plugin-workflow): split dispatch logic to dispatcher
